requirements.txt
django==5.0.6
djangorestframework==3.15.2
pillow==10.4.0
crispy-bootstrap5==2024.2
psycopg2-binary==2.9.9
django-environ==0.11.2